NF: split DeckPickerContextMenuTest
This separate android and non android part
